usage-service/requirements.txt
# Web/API
fastapi
uvicorn[standard]

# Kafka consumer
confluent-kafka 
fastavro

# DB migrations
alembic
psycopg2-binary

# Data processing / MLOps
pandas
apache-airflow
mlflow==2.7.0

# LLM / RAG
langchain
openai
chromadb
sentence-transformers

# Utilities
python-dotenv
requests

#Resilience
pybreaker
tenacity

#Tracing
opentelemetry-api 
opentelemetry-sdk 
opentelemetry-exporter-otlp 
opentelemetry-instrumentation-fastapi 
opentelemetry-instrumentation-requests 
opentelemetry-instrumentation-sqlalchemy
opentelemetry-instrumentation-logging
